requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
httpx==0.25.1
python-dotenv==1.0.0
pydantic==2.4.2
python-multipart==0.0.6
Pillow==10.1.0
pytesseract==0.3.10
PyMuPDF==1.23.5
numpy==1.26.1 